Move ParameterNotUsedInspectionTests for QuickFixes into relevant files under RD.QuickFixes
